Rename config module and remove orchestrator terminology
