+++ pycontextify/chunker.py
@@ -10,7 +10,7 @@ from abc import ABC, abstractmethod
 from pathlib import Path
 from typing import Any, Dict, List
 
-from .orchestrator_config import Config
+from .config import Config
 from .storage_metadata import ChunkMetadata, SourceType
 
 logger = logging.getLogger(__name__)
