+++ pycontextify/indexer.py
@@ -7,6 +7,6 @@ configuration and orchestration utilities.
 from .indexer_loaders import LoaderFactory
 from .indexer_manager import IndexManager
 from .indexer_pdf_loader import PDFLoader
-from .orchestrator_config import Config
+from .config import Config
 
 __all__ = ["IndexManager", "LoaderFactory", "PDFLoader", "Config"]
