+++ tests/integration/test_bootstrap_integration.py
@@ -27,7 +27,7 @@ from unittest import mock
 
 import pytest
 
-from pycontextify.orchestrator_config import Config
+from pycontextify.config import Config
 from pycontextify.indexer_manager import IndexManager
 
 # ============================================================================
