+++ tests/integration/test_persistence.py
@@ -5,7 +5,7 @@ from pathlib import Path
 
 import pytest
 
-from pycontextify.orchestrator_config import Config
+from pycontextify.config import Config
 from pycontextify.indexer_manager import IndexManager
